refactor(modularization): Phase 5 step 5-S2c — evaluation DB carve (task #26)
Move the ``evaluation_v2`` pipeline ORM rows — ``EvaluationDataset`` /
``EvaluationDatasetItem`` / ``EvaluationRun`` / ``EvaluationRunItem``
/ ``EvaluationRunItemAttempt`` — plus their five enums
(``EvaluationDatasetSourceType`` / ``EvaluationRunStatus`` /
``EvaluationRunItemStatus`` / ``EvaluationRunItemAttemptStatus`` /
``EvaluationJudgeMode``) into ``aperag/domains/evaluation/db`` and
extend the aggregate re-export shim in ``aperag.db.models`` with 10
new symbols.

Legacy v1 evaluation tables (``Evaluation`` / ``EvaluationItem`` /
``QuestionSet`` / ``Question`` / ``EvaluationStatus`` /
``EvaluationItemStatus`` / ``QuestionType``) are intentionally NOT
moved — they belong to an older pre-v2 flow that Phase 6 cleanup will
scrub along with the rest of the re-export block.

Snapshot-on-run semantics (value-copied ``input_message`` /
``expected_answer`` / ``tags`` on ``EvaluationRunItem``) travel with
the classes; there is no Python-level cross-domain dep, so the carve
is purely mechanical. ``_random_id`` / ``_enum_column`` helper twins
mirror the 5-S2a / 5-S2b pattern.

``aperag.db.models`` now re-exports 40 symbols post-5-S2c (Phase 3 15
+ Phase 5 conversation 9 + agent_runtime 6 + evaluation 10) — the G11
final lock is the enforcement target for 5-S8.

Task #23 ``_TERMINAL_RUN_STATUSES`` classmethod promotion
(``EvaluationRunStatus.is_terminal()``) is deferred to 5-S6 per
msg=92f5788d Section 6 risk #6; 5-S2c keeps the enum value list
identical to the legacy definition so the repository can keep using
the current ``frozenset`` alias with zero call-site disruption.

Verify:
* ``import aperag.db.models`` + 10 symbol identity check OK
* ``EvaluationDataset is aperag.domains.evaluation.db.models.EvaluationDataset``
* ``pytest tests/unit_test/`` 671 / 29 skipped
* ``uvx ruff check`` + ``uvx ruff format --check`` clean
* ``Base.metadata.tables`` 37 tables still registered
